Add new user struct and its CRUD operations in main.go file
A User struct was added to main.go file with handling of CRUD operations for the user data. The user's data is stored in BadgerDB, a fast key-value DB written in Go. The user's data is stored as JSON and uses uuid for unique keys. Additionally, the go.mod and go.sum files were added to handle use of required packages.