Remove obsolete blob functionality from storage library
The `myblob()` function from `blob.go` has been deleted in this commit. This function was previously responsible for creating a new blob client for Azure's Blob Storage service. However, with recent changes and refactoring of the storage module, this code is no longer necessary. It was removed to declutter the codebase and to ensure that all storage transactions use the updated and streamlined interfaces.