Add version endpoint to API
Added an endpoint '/version' to the API that retrieves and provides the version number of the current build. This aims to help keep track of the software version that is currently in use, thus making easy to identify issues and their corresponding versions for debugging purposes. Implementation details comprise importing new version module and adding new version handler method to the monitoring module.