"Added new route to add data and updated dependencies"
This commit introduces a new route for the addition of data in the application. The change was made because there was previously no route available for users to add new data. Additionally, library dependencies were updated to their latest versions in order to access new features and maintain secure, effective code base. A new view `PutDataView.vue` was also added to handle the new route.