Update README and cryptography keys
This commit updates the README file and changes the array of string keys used within the internal cryptography file. The Readme file changes include detailed tasks for the Web UI application, which is part of the ongoing development scope. This provides more transparency for the development track. Cryptography keys were modified for security reasons to maintain the integrity of the system.